src/python/PythonSDK/requirements.txt
azure-appconfiguration-provider==1.0.0
azure-identity==1.14.0
azure-keyvault-secrets==4.7.0
azure-search-documents==11.4.0
azure-storage-blob==12.18.2
chromadb==0.4.14
duckduckgo-search==3.9.3
flask-sqlalchemy==3.1.1
langchain==0.1.3
langchain-experimental==0.0.49
langchain-openai==0.0.3
numpy==1.26.0
openai==1.9.0
opentelemetry-sdk==1.22.0
pandas==2.1.1
psycopg2==2.9.9
pylint==3.0.2
pymysql==1.1.0
pyodbc==5.0.1
sqlalchemy==2.0.21
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.2
typing==3.7.4.3
unstructured==0.10.22
unstructured[all-docs]==0.10.22
